remove colored, remove file logging and move setup logging to scraper.py
