+++ scraper.py
@@ -1,4 +1,6 @@
 import logging
+import logging.config
+
 from healthtools.scrapers.doctors import DoctorsScraper
 from healthtools.scrapers.foreign_doctors import ForeignDoctorsScraper
 from healthtools.scrapers.clinical_officers import ClinicalOfficersScraper
@@ -6,8 +8,16 @@ from healthtools.scrapers.health_facilities import HealthFacilitiesScraper
 from healthtools.scrapers.nhif_inpatient import NhifInpatientScraper
 from healthtools.scrapers.nhif_outpatient import NhifOutpatientScraper
 from healthtools.scrapers.nhif_outpatient_cs import NhifOutpatientCsScraper
+from healthtools.config import LOGGING
 
-from healthtools.setupLogging import setup_logging
+def setup_logging(default_level=logging.INFO):
+    """
+    Setup logging configuration
+    """
+    try:
+        logging.config.dictConfig(LOGGING)
+    except Exception as ex:
+        logging.basicConfig(level=default_level)
 
 if __name__ == "__main__":
     # set logging
